requirements.txt
asgiref==3.3.1
Django==3.1.6
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.7.0
pytz==2021.1
sqlparse==0.4.1
mysqlclient==2.0.2
django-mysql==2.2.0

numpy~=1.20.3
pandas~=1.2.4
scikit-learn~=0.24.2